requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioprometheus==23.12.0
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argcomplete==3.5.3
async-timeout==5.0.1
attrdict==2.0.1
attrs==24.3.0
audioread==3.0.1
auto_gptq==0.7.1
autoawq==0.2.5
autoawq_kernels==0.0.9
av==14.0.1
bcrypt==4.2.1
bitsandbytes==0.42.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
crcmod==1.7
cryptography==44.0.0
datamodel-code-generator==0.26.5
datasets==3.2.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
easydict==1.13
ecdsa==0.19.0
editdistance==0.8.1
einops==0.8.0
email_validator==2.2.0
et_xmlfile==2.0.0
eva-decord==0.6.1
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
fvcore==0.1.5.post20221221
gekko==1.2.1
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.12.0
gradio_client==1.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
inflect==5.6.2
iopath==0.1.10
isort==5.13.2
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
jj-pytorchvideo==0.1.5
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaldiio==2.18.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
modelscope==1.22.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.22.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.59.8
openpyxl==3.1.5
optimum==1.23.3
orjson==3.10.15
oss2==2.19.1
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
passlib==1.7.4
pathspec==0.12.1
peft==0.14.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyahocorasick==2.1.0
pyarrow==19.0.0
pyasn1==0.6.1
pycorrector==1.1.2
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pypinyin==0.53.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.20
pytorch-wpe==0.0.1
pytz==2024.2
PyYAML==6.0.2
quantile-python==1.1
qwen-vl-utils==0.0.8
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge==1.0.1
rpds-py==0.22.3
rsa==4.9
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
sentence-transformers==3.1.0
sentencepiece==0.2.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
sse-starlette==2.2.1
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
tenacity==9.0.0
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.14
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torch-complex==0.4.4
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.0
transformers-stream-generator==0.0.5
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
websockets==14.2
xinference==1.2.0
xoscar==0.4.6
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zstandard==0.23.0
